delete() will do nothing if all fields provided is neither primary nor unique keys.

+++ .private/scripts/core/Database.php
@@ -602,13 +602,15 @@ class Database {
 
     // Execute queries and accumulate affected rows.
     return array_reduce($keySets, function($result, $keySet) use($table) {
-      // $keySet
-      $res = sprintf('DELETE FROM %s WHERE ', $table) . implode(' AND ', array_keys($keySet));
+      if ( $keySet ) {
+        // $keySet
+        $res = sprintf('DELETE FROM %s WHERE ', $table) . implode(' AND ', array_keys($keySet));
 
-      $res = self::query($res, array_values($keySet));
+        $res = self::query($res, array_values($keySet));
 
-      if ( $res !== FALSE ) {
-        $result+= $res->rowCount();
+        if ( $res !== FALSE ) {
+          $result+= $res->rowCount();
+        }
       }
 
       return $result;
